rename 'pose_dlc' module back to 'choiceworld' for backwards compatibility
